move divider to common ui
