+++ apps/ui/src/Divider.stories.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { Divider, type DividerProps } from '~/ui/Divider';
+import { Divider, type DividerProps } from './Divider';
 
 import type { Meta, StoryObj } from '@storybook/react';
