+++ apps/ui/src/index.ts
@@ -10,3 +10,4 @@ export * from './Placeholder';
 export * from './Toggle';
 export * from './RadioGroup';
 export * from './Combobox';
+export * from './Divider';
